Represent Date using chrono::NaiveDate
Manipulating Date is currently very inefficient, as it needs to be
converted to NaiveDate to get the weekday and perform any calculation.
Since it's not critical to keep it as small as possible for RPC, it just
makes more sense to represent it using NaiveDate.

Also move weekday_set() as a function of DateRange.