Add an actual Schedule and expose through CLI
Compute a schedule based on the list of timeslots. In fact there's no
need to expose this through RPC: nothing prevents the client from doing
it itself, and that's arguably a better approach (less load on the
server).

The formatting could certainly be improved and it should be made
configurable, but it's already pretty usable!