Actuator thread (event loop)
Bring the Actuator struct to life by giving it a thread, which is
responsible for handling the timeslots themselves. It is basically an
event loop, handling timeslot (and therefore actuator state) changes,
and sleeping until the next change occurs. Communication between the RPC
calls and the actuator thread is done mainly via an ActiveTimeSlot
struct, giving the actuator thread

When timeslots are modified, the thread may need to be woken up early.
The new ActiveTimeSlot struct (and the associated condition variable)
allows the Actuator methods called via RPC to wake up the thread if
necessary and inform it of the new active timeslot. If no RPC occurs,
the actuator thread itself finds the next active timeslot.

The ActiveTimeSlot approach is probably overkill, and finding whether
the thread should be woken up gets quite complicated. In practice,
always waking up the thread should be good enough, but the resource
sharing would need to be refactored to avoid unnecessary context
switches.