Server config file
Configuring the server via the command line isn't going to scale, so
let's use a config file. Go the easy route by deserialising the file
using serde. YAML seems to be a decent choice.

There isn't a real need to modify the list of actuators at runtime, so
let's remove the corresponding API and use a Vec instead of a BTreeMap.
This will allow making the list of actuators read-only, and should make
it easy to merge Server and RpcServer.